Replace interrupt bool with uint32 (atomic)
The interrupt bool is checked both inside and outside of a lock during
writes, resulting in a data race. This change replaces the bool with
a uint32 and adds an accessor method, interrupted(), that returns
whether an interrupt has been set (value irrelevant to this since it's
only ever read/written after acquiring a lock).

The interrupt is now signalled by flipping the interrupt integer to
a non-zero value. This can only be done via atomic methods -- all reads
and writes to that integer must go through the atomic package.