bugfix: this being wrong when rest arguments are used
During a previous fix it seems that rest parameters weren't taken into
consideration and currently `this` refers to them in the provided test.

The test in question is a very simplified variation of a real code.

It seems all of the following things are relevant for the bug to occur:
1. method with only a rest parameter
2. usage of `this`
3. usage of the parameter
4. All of the above within a lambda inside the method